Added Images For Puzzle Sprites
Wasn't sure what would be more useful, so I made a few different
versions for multiplication and increment: alpha w/ and w/o text; and
colorized w/ and w/o text. the piece on the upper left is the given, and
the rest are what the player gets to choose from. For multiplication:
the correct answers are the three pieces to the right of the yellow
piece. For increment: the correct answers are the four pieces to the
right of the yellow piece. pay attention when using the sliced images so
you don't accidently use the wrong one... i can't seem to rename the
slices...